requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.4.24
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
av==14.0.1
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
boto3==1.35.95
botocore==1.35.95
braceexpand==0.1.7
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dacite==1.7.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
easydict==1.13
einops==0.8.0
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpeg==1.4
fiftyone==1.2.0
fiftyone-brain==0.18.2
fiftyone_db==1.1.7
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
glob2==0.7
graphql-core==3.2.5
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
inflate64==1.0.1
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaleido==0.2.1
keyring==25.6.0
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
lightning-utilities==0.11.9
lpips==0.1.4
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mongoengine==0.29.1
more-itertools==10.5.0
motor==3.6.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.1.0
multivolumefile==0.2.3
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
notebook_shim==0.2.4
numpy==1.25.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pipdeptree==2.16.2
pkginfo==1.12.0
platformdirs==4.3.6
plotly==5.24.1
pprintpp==0.4.0
priority==2.0.0
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py7zr==0.22.0
pybcj==1.0.3
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydash==8.0.4
Pygments==2.19.1
pymongo==4.9.2
pyparsing==3.2.1
pyppmd==1.1.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==3.2.1
python-multipart==0.0.20
python-socketio==5.12.1
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.0
pyzstd==0.16.2
rarfile==4.2
readme_renderer==44.0
redis==5.2.1
referencing==0.35.1
reflex==0.6.8
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.32
reflex-pyplot==0.1.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
Rtree==1.3.0
s3transfer==0.10.4
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
sqlmodel==0.0.22
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.41.3
starlette-admin==0.14.1
strawberry-graphql==0.256.1
stringzilla==3.11.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tinycss2==1.4.0
tomlkit==0.13.2
torch==2.5.1
torchinfo==1.8.0
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
twine==6.0.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.59
ultralytics-thop==2.0.13
universal-analytics-python3==1.1.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
voxel51-eta==0.13.1
wavemix==0.2.4
wcwidth==0.2.13
webcolors==24.11.1
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.18.3
zipp==3.21.0
